+++ src/lib/Menu/MenuLinks.ts
@@ -1,4 +1,4 @@
-import {type Action, type ActionIcon, DefaultAction} from '$lib/actions';
+import { type Action, type ActionIcon, type ActionOptions, DefaultAction } from '$lib/actions';
 import type { Optional } from '$lib/genericTypes';
 
 /** */
@@ -12,8 +12,13 @@ export class Submenu extends DefaultAction {
 		return this._links;
 	}
 
-	constructor(label: string, icon: Optional<ActionIcon>, links: Array<MenuLink>) {
-		super(label, icon);
+	constructor(
+		label: string,
+		icon: Optional<ActionIcon>,
+		links: Array<MenuLink>,
+		options?: ActionOptions
+	) {
+		super(label, icon, options);
 		this._links = links;
 	}
 }